Fully refactored subprocess handling on all platforms.
Now we have all process-related code in utils.process, which itself
imports from platform-specific files as needed.

On posix, we have reliable asynchronous delivery of stdout and stderr,
and on win32 at least we have the basics that subprocess.py provides,
since pexpect is not available.

We also now support robust killing of subprocesses that may capture
SIGINT: one SIGINT on our end is sent to the subprocess, but then we
kill it, to prevent a rogue subprocess from hijacking the ipython
console.

Note that on posix, we now depend on pexpect, but we ship our own copy
to users which we'll use if there's no system pexpect installed.

UNC path handling for windows was implemented as a context manager
called AvoidUNCPath.