update parallel code for py3k
This is primarily tweaks of bytes/unicode, but other fixes include:

* some integer division
* added co_kwonlyargcount to code objects
* a few places to handle map/range being objects

non-copying numpy is disabled on py3k, because arrays are not reconstructed properly on the other side. This is because pyzmq
always receives a byte array (memoryview.itemsize=1), which confuses numpy.

With these changes, almost all parallel tests pass.  Notable exceptions are: sync_imports, and MUX engine death.