Avoid creating default connection dir if not needed
If a connection file is specified with an absolute path via the '-f'
command line option, ipykernel still makes sure that the default connection
directory exists, even though it is not needed. Instead, simply create the
directory for the path to the connection file that we end up actually
using.

This introduces a slight change in behaviour in that if one specifies a
connection file via the '-f' option and the given directory does not
exist, it will be created instead of an error being generated.