Import hooks to set up inline backend when matplotlib is loaded
Addresses ipython/ipython#6424

We believe that most people using matplotlib from a Jupyter frontend
want plots to appear inline. Particularly when introducing new users to
Python, we end up telling them to run '%matplotlib inline' with a "don't
worry about what this does...", which feels wrong when we're trying to
help them understand what they're doing.

This is an attempt to make inline plots the default when you use
matplotlib in an IPython kernel. It installs import hooks to watch for
matplotlib being imported, and sets up the inline backend when it is.
There is a config option (IPKernelApp.mpl_plots_inline) to disable this
behaviour. Specifying a backend with %matplotlib or %pylab should also
override it.

We've discussed before whether this belongs in IPython or in matplotlib.
I don't have a strong preference, but I got the impression that other
people preferred it to be in IPython. If we decide to go the other way,
I'm happy to help mpl do it.