Don't override default display name with default kernel name
I installed the Python 2 kernel spec with:

python2 -m ipykernel.kernelspec --user

I noticed that the display name was showing up as 'python2' instead of
'Python 2'.

The name flag in the command line arguments now has an explicit default
of KERNEL_NAME. This gets passed to install(), which expects None to
mean the default. It treats it as a non-default name and sets the
display name as well.

This changes the logic of install() so that the display name is not set
to the default KERNEL_NAME, and it shows up as 'Python 2' again.