Add a launcher module to ignore the CWD for imports while ipykernel starts
After helping someone who'd shadowed a stdlib module with a file in the
cwd (jupyter/help#97), I realised that we could fairly easily make the
kernel able to start in such situations.

This PR adds an ipykernel_launcher module which removes the CWD from
sys.path and then launches the kernel. There is already code in IPython
which adds the CWD back to sys.path before user code runs.

This doesn't get rid of issues with name clashes: modules that we import
will shadow modules of the same name in the cwd. So if you wrote
math.py, you wouldn't be able to import it. But I think this is easier
for users to understand than the kernel completely failing to start,
especially if it's not easy for them to see the logs showing the error.