Add and register a comm to register *other* comms
This comm allows a client to send comm-messages to it, containing
data specifying a new comm to register on the kernel. This minimal
comm allows much greater extension of the kernel's behaviour, when
access to the kernel object isn't available.

The comm is registered, but not started, when the kernel is
initialized. A comm-open message can be sent to start it, then
comm-messages sent to register new comms, to perform arbitrary
actions on the kernel side.