Replace function for in-process kernel compatibility
From v5.0.0, the default IPythonKernel contains an _abort_queues function invoked when a shell execute request generates an exception. This schedules a message, which clears then clears the abort. This process does not work in the InProcessKernel. After an exception, all shell execute requests return an "ERROR: execution aborted" message. This is fixed by overriding the _abort_queues function in the InProcessKernel.