Add new trio_loop config flag that runs Trio on main thread
This flag causes Trio to run on the main thread, and puts ZMQ and other
stuff on a background thread. When trio runs on the main thread, async
cells are executed as tasks inside a global nursery. A global nursery
named GLOBAL_NURSERY is exported in builtins so that background tasks
can continue to run even after a cell finishes executing. Exceptions in
background tasks are caught and displayed in the current cell.

This implicitly runs `%autoawait trio` when the kernel starts and then
disables the %autoawait magic so that users can't switch to a different
loop while in Trio loop mode.

Co-authored-by: Brian Mackintosh <bcmackintosh@gmail.com>