Some removal of ipython_genutils.py3compat.
As in some of the other Jupyter Repositories; this start to remove
ipython_genutils because of packaging problem in linux distribution
(getting rid of nose).

This is a step toward that.

Note that this changes is slightly stricter in some area where function
would previously accept string/bytes and now only accept one; but as far
as I can tell are generally only used with one of those since we are
Python 3 only.