IPython's start_kernel is deprecated.
Technically it has been deprecated since 5.0 but indirectly as it
imports from IPython.kernel with itself is a shim module that exposes
ipykernel.

So it is useless to call into IPython if it's for calling back into
ipykernel.

Though the start_kernel itself was not emitting a deprecation warning,
which will be the case in 8.0, and hopefully we can finish removing the
shim modules later.