Try to pass cell id to executing kernel.
Is seem that few kernel want to make use of cell id for various
reasons.

One way to get the cell_id in the executing context is to make use of
init_metadata but it is marked for removal.

for example in [there](https://github.com/robots-from-jupyter/robotkernel/blob/19b28267406d1f8555ce73b96e7efb8af8417266/src/robotkernel/kernel.py#L196-L205)

Another is to start passing cell_id down.

This is technically a change of API as our consumer may not support more
parameters so we take a peak at the signature, and pass cell_id only if
it is :
- an explicit parameter,
- the function/method takes varkwargs.

We could also start to refactor to pass a metadata object with multiple
fields if this is the preferred route.

One questions is whether and how we want to deprecated not receiving
cell_id as a parameter.

Related to https://github.com/ipython/ipython/issues/13579 and
subsequent PR on IPython side to support cell_id in progress.