Configurables needs to be configurable.
IPython rely on configurables being configurables to ... configure
things.

Comm seem to insert or return a non-configurable in IPython, that make
ipykernel crash in some instances. For example but not limited to:

    %config Completer.use_jedi = False

Or trying to use hvplot for example.

This is just a "this is wrong" PR, there should be a proper fix,
and I will likely add a hard check in IPython that what is added in
configurables is actually a configurables.

This will need a fix in a Comm pacakge.