Add unit test for a connection file with ports set to 0
Before ipykernel 6.23.3, i.e., before https://github.com/ipython/ipykernel/pull/1127, a kernel manager could specify a channel port of 0, and ipykernel would pick a random port and rewrite the connection file with the actual port used. This provided a nice way to address the natural race condition between a kernel manager picking a port and ipykernel actually connecting to it and using it.

This unit test tests that this port 0 connection file behavior works, and also tests that existing information in the connection file is not overwritten.