Detect parent change in more cases on unix
New behaviour: poll getppid and exit if this value is no longer equal to
the original parent pid.

For backwards compatibility: keep old behaviour when parent_pid is 0 or
not given.

Parentpoller only watches for parent change to pid 1 (init), but in
modern linux environments processes are usually reparented to the
systemd user session process. In this situation, ipykernel never
noticies that its parent has exited.

jupyter-client already sets JPY_PARENT_PID with the pid of the parent
process. We could trust this environment variable directly, but in this
change, it is only used if it is equal to the actual parent id when the
poller starts. (A side effect: if the parent dies quickly enough,
ipykernel does not notice.)