Try to fix unclosed resources.
I think these might be one of the few places that create those socket,
so maybe closing them is the right call. I'm not sure.

There are random failures on some tests.

as _handle_pipe_msg is use in a single place and `self._async_pipe_in1`
is used only to be passed between _handle_pipe_msgs and _handle_pipe_msg
I'm tempted to refactor and pass _async_pipe_in1 as a parameter maybe ?
and also rename to _handle_single_pipe_msg to make it visually a bit
more distinctive.