Remove test_check job
Same as #1318, there is no reasons for it to be there anymore, and it is more
annoying than anything else. It used to be usefull if you wanted to enforce some
CI items but not other, but this is now native via branch protection rules, but
now Github UI shows which jobs are required to pass for a PR to be merged, and
with this it does not.