Remove deprecated modules since 4.3 (2016).
This is annoying as this makes downstream tests fail in IPython,
while the culprit is ipyparallel.

Ipyparallel itself is tested, and main has been fixed, so we just get
a false failure in IPython downstream tests.

It's also been deprecated for 10 years, now, so at least we can have
that removed from main; so that we can see if downstream breaks.