TQDM workaround due to unresponsive maintainer
tqdm has an incorrect detection of
ZMQInteractiveShell when launch via a scheduler
that bypass IPKernelApp. Think of JupyterHub
cluster spawners and co. As of end of Feb 2025,
the maintainer has been unresponsive for 5 months,
to our fix, so we implement a workaround. I don't
like it but we have few other choices. See
`https://github.com/tqdm/tqdm/pull/1628`